Rename routes.py to the more accurate route_handlers.py
